+++ layout_skripsi/kategori/read_event_creator.php
@@ -63,7 +63,7 @@ $eksekusi = mysqli_query($conn, $sql);
                 <a href="form_edit_event_creator.php?id_event_creator=<?php echo $row['id_event_creator'] ?>" class="w3-button w3-border w3-small w3-blue"> Edit </a> 
               </td>
               <td>
-                <a href="hapus.php?id_event_creator=<?php echo $row['id_event_creator'] ?>"  class="w3-button w3-border w3-small w3-red"> Hapus </a>
+                <a href="hapus.php?id_event_creator=<?php echo $row['id_event_creator'] ?>" onclick="return confirm('Anda yakin mau menghapus item ini ?')" class="w3-button w3-border w3-small w3-red"> Hapus </a>
               </td>
               <td>
                 <a href="hapus.php?id_kategori_event=<?php echo $row['id_kategori_event'] ?>"  class="w3-button w3-border w3-small w3-deep-purple"> Lihat Detail </a>
